add support for creating WHERE clauses
The original plan for this was to use the python comprehension syntax.
However, that would be too much effort, and for efficiency I'd want to modify
the bytecode of the generators. Not fun.

Instead, I've gone for setting up a thin dialect based on comparison operators, and
binary operators for the 'AND' and 'OR'.

This also adds basic support for dates, integers, and strings for query values, as well
as a thin escaping mechanism (which is not a substitute for protecting against SQL injection).